New Fixes on fileUpload
